+++ cran_gen/volume.h
@@ -25,8 +25,11 @@
 #include <boost/random/normal_distribution.hpp>
 #include <boost/random/uniform_real_distribution.hpp>
 #include "vars.h"
-#include "polytopes.h"
+#include "hpolytope.h"
+#include "vpolytope.h"
+#include "zpolytope.h"
 //#include "ellipsoids.h"
+#include "ball.h"
 #include "ballintersectconvex.h"
 #include "vpolyintersectvpoly.h"
 #include "samplers.h"