+++ include/cartesian_geom/point.h
@@ -11,7 +11,7 @@
 
 #include <iostream>
 
-template <class K>
+template <typename K>
 class point
 {
 private:
@@ -24,29 +24,33 @@ public:
 
     point() {}
     
-    point(const unsigned int dim) {
+    point(const unsigned int &dim) {
         d = dim;
         coeffs = Coeff(d,0);
     }
     
-    point(const unsigned int dim, iter begin, iter endit) {
+    point(const unsigned int &dim, iter begin, iter endit) {
         d = dim;
         coeffs = Coeff(begin,endit);
     }
     
-    int dimension() {
+    int dimension() const {
         return d;
     }
     
-    void set_dimension(const unsigned int dim) {
+    void set_dimension(const unsigned int &dim) {
         d = dim;
     }
     
-    void set_coord(const unsigned int i, FT coord) {
+    void set_coord(const unsigned int &i, const FT &coord) {
         coeffs[i] = coord;
     }
+
+    Coeff get_coeffs() {
+        return coeffs;
+    }
     
-    FT operator[] (const unsigned int i) {
+    FT operator[] (const unsigned int &i) const {
         return coeffs[i];
     }
     
@@ -153,8 +157,8 @@ public:
     
 };
 
-template<class K>
-point<K> operator* (const typename K::FT& k, point<K>& p) {
+template<typename K>
+point<K> operator* (typename K::FT const& k, point<K>& p) {
     return p * k;
 }