+++ include/volume/vars.h
@@ -13,7 +13,7 @@
 #define VARS_H
 
 //structs with variables and random generators
-template <typename NT, class RNG>
+template <typename NT, typename RNG>
 struct vars{
 public:
     typedef RNG RNGType;
@@ -27,6 +27,7 @@ public:
           NT up,
           const int L,
           NT che_rad,
+          NT diameter,
           RNG &rng,
           boost::random::uniform_real_distribution<>(urdist),
           boost::random::uniform_real_distribution<> urdist1,
@@ -41,7 +42,7 @@ public:
           bool rdhr_walk
     ) :
             m(m), n(n), walk_steps(walk_steps), n_threads(n_threads), err(err), error(error),
-            lw(lw), up(up), L(L), che_rad(che_rad), rng(rng),
+            lw(lw), up(up), L(L), che_rad(che_rad), diameter(diameter), rng(rng),
             urdist(urdist), urdist1(urdist1) , delta(delta) , verbose(verbose), rand_only(rand_only), round(round),
             NN(NN),birk(birk), ball_walk(ball_walk), cdhr_walk(cdhr_walk), rdhr_walk(rdhr_walk){};
 
@@ -55,6 +56,7 @@ public:
     NT up;
     const int L;
     NT che_rad;
+    NT diameter;
     RNG &rng;
     boost::random::uniform_real_distribution<>(urdist);
     boost::random::uniform_real_distribution<> urdist1;
@@ -69,7 +71,7 @@ public:
     bool rdhr_walk;
 };
 
-template <typename NT, class RNG>
+template <typename NT, typename RNG>
 struct vars_g{
 public:
     typedef RNG RNGType;
@@ -122,5 +124,37 @@ public:
     bool cdhr_walk;
     bool rdhr_walk;
 };
+
+
+template <typename NT>
+struct vars_ban {
+public:
+
+    vars_ban(NT lb,
+             NT ub,
+             NT p,
+             NT rmax,
+             NT alpha,
+             int win_len,
+             int N,
+             int nu,
+             bool window2
+    ) :
+            lb(lb), ub(ub), p(p), rmax(rmax), alpha(alpha),
+            win_len(win_len), N(N), nu(nu), window2(window2) {};
+
+
+    NT lb;
+    NT ub;
+    NT p;
+    NT rmax;
+    NT alpha;
+    int win_len;
+    int N;
+    int nu;
+    bool window2;
+};
+
+
 
 #endif