+++ include/volume/volume.h
@@ -24,8 +24,10 @@
 #include "random/normal_distribution.hpp"
 #include "random/uniform_real_distribution.hpp"
 #include "vars.h"
-#include "polytopes.h"
-//#include "ellipsoids.h"
+#include "hpolytope.h"
+#include "vpolytope.h"
+#include "zpolytope.h"
+#include "ball.h"
 #include "ballintersectconvex.h"
 #include "vpolyintersectvpoly.h"
 #include "samplers.h"
@@ -42,7 +44,7 @@
 //#include "copulas.h"
 
 
-template <class Polytope, class Parameters, class Point, typename NT>
+template <typename Polytope, typename Parameters, typename Point, typename NT>
 NT volume(Polytope &P,
           Parameters & var,  // constans for volume
           std::pair<Point,NT> InnerBall)  //Chebychev ball
@@ -254,6 +256,7 @@ NT volume(Polytope &P,
     if(print) std::cout<<"volume computed: "<<vol<<std::endl;
     #endif
 
+    P.free_them_all();
     return vol;
 }
 
@@ -262,7 +265,7 @@ NT volume(Polytope &P,
 // Implementation is based on algorithm from paper "A practical volume algorithm",
 // Springer-Verlag Berlin Heidelberg and The Mathematical Programming Society 2015
 // Ben Cousins, Santosh Vempala
-template <class Polytope, class UParameters, class GParameters, class Point, typename NT>
+template <typename Polytope, typename UParameters, typename GParameters, typename Point, typename NT>
 NT volume_gaussian_annealing(Polytope &P,
                              GParameters & var,  // constans for volume
                              UParameters & var2,
@@ -439,6 +442,7 @@ NT volume_gaussian_annealing(Polytope &P,
     }
     #endif
 
+    P.free_them_all();
     return vol;
 }